requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langgraph==0.2.19
ragatouille==0.0.8.post4
chromadb==0.5.5
openai==1.43.0
tiktoken==0.7.0
transformers==4.44.2
torch==2.4.1
python-dotenv==1.0.1
unstructured[all-docs]==0.15.9